requirements.txt
# LangChain Core
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.3.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.2.0

# Vector Store
chromadb>=0.5.0

# Telegram
python-telegram-bot>=21.0

# Web
fastapi
uvicorn
jinja2
python-multipart

# Utilities
requests
beautifulsoup4
psutil
paramiko
pyyaml
dateparser
pytz
ddgs
youtube-transcript-api
python-dateutil
